+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Added Clone implementation for 'static events.
 - On Windows, fix window intermittently hanging when `ControlFlow` was set to `Poll`.
 - On Windows, fix `WindowBuilder::with_maximized` being ignored.
 - On Android, minimal platform support.